This project creates wrapper classes for use with OSISoft's AFSDK.  The classes expose some of the AFSDK functionality to a COM application.
You should be able to copy the AFWrapper.dll and AFWrapper.tlb files to your workstation and reference them by navigating to the .tlb file with your VBA tools Reference dialog.

There are many more AFSDK methods that could be exposed by this wrapper.  Anyone interested in doing this is asked to please create a new branch to add the desired functionality and merge it back into the project when complete.